Move playwrite stuff to utils folder
